py/requirements.txt
aiobotocore==0.9.2
aiodns==1.1.1
aiohttp==3.3.1
aiohttp-session==2.5.1
arq==0.13
asyncpg==0.15.0
async-timeout==3.0.0
bcrypt==3.1.4
buildpg==0.0.2
cchardet==2.1.1
cryptography==2.2.2
google-auth==1.4.1
lorem==0.1.1
pydantic[email]==0.11.2
Pillow==5.1.0
raven==6.8.0
uvloop==0.9.1
